Splitting things out for easy testing
