+++ backend/server/public/package.json
@@ -4,15 +4,17 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "build": "webpack --progress",
-    "watch": "webpack -w --progress"
+    "watch": "webpack -w --progress",
+    "test:watch": "npm test -- --watch"
   },
   "author": "Jeff Larson <jeff.larson@propublica.org>",
   "license": "MIT",
   "devDependencies": {
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.2.1",
+    "babel-jest": "^22.1.0",
     "babel-loader": "^7.1.2",
     "babel-plugin-jsx-pragmatic": "^1.0.2",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
@@ -27,6 +29,7 @@
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.6",
     "html-webpack-plugin": "^2.30.1",
+    "jest": "^22.1.4",
     "node-sass": "^4.7.2",
     "sass-loader": "^6.0.6",
     "source-map-loader": "^0.2.3",
