+++ backend/server/public/src/actions.js
@@ -1,0 +1,185 @@
+import { headers } from "utils.js";
+
+export const NEW_ADS = "new_ads";
+export const newAds = ads => ({
+  type: NEW_ADS,
+  value: ads
+});
+
+export const GOT_THAT_AD = "GOT_THAT_AD";
+export const receiveOneAd = ad => ({
+  type: GOT_THAT_AD,
+  ad: ad
+});
+
+export const REQUESTING_ONE_AD = "REQUESTING_ONE_AD";
+export const requestingOneAd = ad_id => ({
+  type: REQUESTING_ONE_AD,
+  ad_id: ad_id
+});
+
+export const SET_LANG = "set_lang";
+export const setLang = lang => ({
+  type: SET_LANG,
+  value: lang
+});
+
+export const NEW_SEARCH = "new_search";
+export const newSearch = query => ({
+  type: NEW_SEARCH,
+  value: query
+});
+
+export const BATCH = "batch";
+export const batch = (...actions) => {
+  return {
+    type: BATCH,
+    actions: actions
+  };
+};
+
+const a = type => arg => ({ type, value: arg });
+export const NEW_ENTITIES = "new_entities";
+export const NEW_ADVERTISERS = "new_advertisers";
+export const NEW_TARGETS = "new_targets";
+export const newEntities = a(NEW_ENTITIES);
+export const newAdvertisers = a(NEW_ADVERTISERS);
+export const newTargets = a(NEW_TARGETS);
+
+export const FILTER_ENTITY = "filter_entity";
+export const FILTER_ADVERTISER = "filter_advertiser";
+export const FILTER_TARGET = "filter_target";
+export const filterEntity = a(FILTER_ENTITY);
+export const filterAdvertiser = a(FILTER_ADVERTISER);
+export const filterTarget = a(FILTER_TARGET);
+
+export const TOGGLE_TARGET = "toggle_target";
+export const TOGGLE_ADVERTISER = "toggle_advertiser";
+export const TOGGLE_ENTITY = "toggle_entity";
+export const RESET_DROPDOWNS = "reset_dropdowns";
+export const toggleTarget = () => ({ type: TOGGLE_TARGET });
+export const toggleAdvertiser = () => ({ type: TOGGLE_ADVERTISER });
+export const toggleEntity = () => ({ type: TOGGLE_ENTITY });
+export const resetDropdowns = () => ({ type: RESET_DROPDOWNS });
+
+export const NEXT_PAGE = "next_page";
+export const PREV_PAGE = "prev_page";
+export const SET_PAGE = "set_page";
+export const SET_TOTAL = "set_total";
+export const nextPage = () => ({ type: NEXT_PAGE });
+export const prevPage = () => ({ type: PREV_PAGE });
+export const setPage = page => ({ type: SET_PAGE, value: page });
+export const setTotal = total => ({ type: SET_TOTAL, value: total });
+
+export const getOneAd = (ad_id, url = "/facebook-ads/ads") => {
+  if (!ad_id) return () => null;
+
+  let path = `${url}/${ad_id}`;
+  return (dispatch, getState) => {
+    let state = getState();
+
+    if (
+      state.permalinked_ad &&
+      state.permalinked_ad.ads &&
+      state.permalinked_ad.ads[ad_id] &&
+      state.permalinked_ad.ads[ad_id].id
+    ) {
+      return Promise.resolve(
+        dispatch(receiveOneAd({ ads: [state.permalinked_ad.ads[ad_id]] }))
+      );
+    }
+    dispatch(requestingOneAd(ad_id));
+
+    fetch(path, {
+      method: "GET",
+      headers: headers(state.credentials, state.lang)
+    })
+      .then(res => res.json())
+      .then(ad => {
+        dispatch(receiveOneAd(ad));
+      });
+  };
+};
+
+// Admin actions
+export const HIDE_AD = "hide_ad";
+export const LOGIN = "login";
+export const LOGOUT = "logout";
+export const hideAd = ad => ({
+  type: HIDE_AD,
+  id: ad.id
+});
+
+export const suppressAd = ad => {
+  return (dispatch, getState) => {
+    dispatch(hideAd(ad));
+    return fetch("/facebook-ads/admin/ads", {
+      method: "POST",
+      body: ad.id,
+      headers: headers(getState().credentials)
+    }).then(resp => {
+      if (resp.ok) {
+        console.log("suppressed");
+      } else {
+        dispatch(logout());
+      }
+    });
+  };
+};
+
+export const login = credentials => ({
+  type: LOGIN,
+  value: credentials
+});
+
+export const logout = () => ({
+  type: LOGOUT
+});
+
+export const b64 = thing =>
+  btoa(thing)
+    .replace(/\+/g, "-")
+    .replace(/\//g, "_")
+    .replace(/=/g, "");
+
+export const createJWT = (username, password) => {
+  const encoder = new TextEncoder();
+  const header = {
+    alg: "HS256",
+    typ: "JWT"
+  };
+  const payload = {
+    username
+  };
+  const base = `${b64(JSON.stringify(header))}.${b64(JSON.stringify(payload))}`;
+  const encoded = encoder.encode(base);
+  return window.crypto.subtle
+    .importKey(
+      "raw",
+      encoder.encode(password),
+      { name: "HMAC", hash: { name: "SHA-256" } },
+      false,
+      ["sign"]
+    )
+    .then(key => window.crypto.subtle.sign({ name: "HMAC" }, key, encoded))
+    .then(signature => ({
+      token: `${base}.${b64(
+        String.fromCharCode.apply(null, new Uint8Array(signature))
+      )}`
+    }));
+};
+
+export const authorize = (username, password) => {
+  // create jwt
+  return dispatch =>
+    createJWT(username, password).then(token => {
+      return fetch("/facebook-ads/login", {
+        method: "POST",
+        headers: headers(token)
+      }).then(resp => {
+        if (resp.ok) {
+          dispatch(login(token));
+        }
+      });
+    });
+};
