+++ backend/server/public/src/components/ad.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Targeting from "./targeting.jsx";
+
+const Ad = ({ ad }) => (
+  <div>
+    <div className="message">
+      <div dangerouslySetInnerHTML={{ __html: ad.html }} />
+    </div>
+    {ad.targeting !== null ? <Targeting targeting={ad.targeting} /> : ""}
+  </div>
+);
+export default Ad;
