+++ backend/server/public/src/utils.js
@@ -1,16 +1,18 @@
 import "url-search-params-polyfill";
-import { setTotal, setPage } from "pagination.jsx";
 import {
+  batch,
+  setLang,
+  setTotal,
+  setPage,
+  newAds,
   newAdvertisers,
   newEntities,
+  newSearch,
   newTargets,
-  serializeAdvertisers,
-  serializeTargets,
-  serializeEntities,
   filterAdvertiser,
   filterEntity,
   filterTarget
-} from "filters.jsx";
+} from "actions.js";
 
 const auth = credentials =>
   credentials ? { Authorization: `Bearer ${credentials.token}` } : {};
@@ -20,53 +22,34 @@ const headers = (credentials, lang) =>
 
 const language = lang => ({ "Accept-Language": lang + ";q=1.0" });
 
-const NEW_ADS = "new_ads";
-const newAds = ads => ({
-  type: NEW_ADS,
-  value: ads
-});
-
-const GOT_THAT_AD = "GOT_THAT_AD";
-const receiveOneAd = ad => ({
-  type: GOT_THAT_AD,
-  ad: ad
-});
-
-const REQUESTING_ONE_AD = "REQUESTING_ONE_AD";
-const requestingOneAd = ad_id => ({
-  type: REQUESTING_ONE_AD,
-  ad_id: ad_id
-});
-
-const SET_LANG = "set_lang";
-const setLang = lang => ({
-  type: SET_LANG,
-  value: lang
-});
-
-const lang = (state = null, action) => {
-  switch (action.type) {
-    case SET_LANG:
-      return action.value;
-    default:
-      return state;
-  }
+// https://github.com/reactjs/redux/issues/911#issuecomment-149192251
+export const enableBatching = reducer => {
+  return function batchingReducer(state, action) {
+    switch (action.type) {
+      case BATCH:
+        return action.actions.reduce(batchingReducer, state);
+      default:
+        return reducer(state, action);
+    }
+  };
 };
 
-const NEW_SEARCH = "new_search";
-const newSearch = query => ({
-  type: NEW_SEARCH,
-  value: query
-});
-
-const search = (state = null, action) => {
-  switch (action.type) {
-    case NEW_SEARCH:
-      return action.value;
-    default:
-      return state;
-  }
+const s = (plural, singular, map) => {
+  return (params, state) => {
+    if (!state[plural]) return params;
+    const items = state[plural]
+      .filter(it => it.active)
+      .map(it => (map ? map(it[singular]) : it[singular]));
+    if (items.length > 0) {
+      params.set(plural, JSON.stringify(items));
+    }
+    return params;
+  };
 };
+
+const serializeEntities = s("entities", "entity", entity => ({ entity }));
+const serializeAdvertisers = s("advertisers", "advertiser");
+const serializeTargets = s("targets", "target", target => ({ target }));
 
 const serialize = store => {
   let params = new URLSearchParams();
@@ -135,57 +118,6 @@ const deserialize = dispatch => {
   dispatch(batch(...actions));
 };
 
-const BATCH = "batch";
-const batch = (...actions) => {
-  return {
-    type: BATCH,
-    actions: actions
-  };
-};
-
-// https://github.com/reactjs/redux/issues/911#issuecomment-149192251
-const enableBatching = reducer => {
-  return function batchingReducer(state, action) {
-    switch (action.type) {
-      case BATCH:
-        return action.actions.reduce(batchingReducer, state);
-      default:
-        return reducer(state, action);
-    }
-  };
-};
-
-// getOneAd() and refresh() are thunk action creators.
-const getOneAd = (ad_id, url = "/facebook-ads/ads") => {
-  if (!ad_id) return () => null;
-
-  let path = `${url}/${ad_id}`;
-  return (dispatch, getState) => {
-    let state = getState();
-
-    if (
-      state.permalinked_ad &&
-      state.permalinked_ad.ads &&
-      state.permalinked_ad.ads[ad_id] &&
-      state.permalinked_ad.ads[ad_id].id
-    ) {
-      return Promise.resolve(
-        dispatch(receiveOneAd({ ads: [state.permalinked_ad.ads[ad_id]] }))
-      );
-    }
-    dispatch(requestingOneAd(ad_id));
-
-    fetch(path, {
-      method: "GET",
-      headers: headers(state.credentials, state.lang)
-    })
-      .then(res => res.json())
-      .then(ad => {
-        dispatch(receiveOneAd(ad));
-      });
-  };
-};
-
 // this is horrid, todo cleanup
 let loaded = false;
 const refresh = (store, url = "/facebook-ads/ads") => {
@@ -228,17 +160,4 @@ const refresh = (store, url = "/facebook-ads/ads") => {
   }
 };
 
-export {
-  headers,
-  newAds,
-  NEW_ADS,
-  search,
-  getOneAd,
-  GOT_THAT_AD,
-  REQUESTING_ONE_AD,
-  refresh,
-  newSearch,
-  deserialize,
-  enableBatching,
-  lang
-};
+export { headers, refresh, deserialize };
